safely check for pymc3 objects without dependency
Without introducing a fixed dependency on `pymc3`
(or any other non-standard library module),
this commit introduces facilities to safely
recognize `pymc3` objects (and objects with
origin in other modules).
This is achieved by reliance on standard
library module `importlib`.
In summary, we add a function with signature
`_has_type(object_, typename, module_path)`
that checks if `object_` has a type given as
string `typename` that can be found under a
`module_path`, also given as string.
If a module under `module_path` can not be imported
(e.g. because it is not installed), we returna `False`
`to cover cases where a user
does not use `pymc3`. We use this function to
refactor a few unsafe string-based checks
in `arviz.utils.utils`, for further information
on those check issue #138.