Added orderedDict to forestplot (#618)
* Added orderedDict to forestplot

I added OrderedDict module from collections to keep the structure of the InferenceData object. This applies to both the variable names and the dimensions within the variable names. Fixes #613

* reintroduced set().union for var_names if models > 2

I had to reintroduce set().union when var_names are not specified by the user and when there is more than one model.

* Ran black

* Update forestplot.py

* Update forestplot.py

* Pylint fixes

Note that I had to turn the summary of forestplot() into one sentence (replacing dot by a comma).

* OrderedDict also fixes #614

Adding another OrderedDict() instead of {} fixes #614

* Pylint fix

Removed previous created comma for line 42 and just created a single line for summary of forestplot()

* Test short docstring

* fix whitespace