Add contours of highest posterior density for 2D KDE plot (#1665)
* Add a function to calculate HPD regions on 2D KDEs

This commit adds a function `_find_hpd_contours` that is designed to
calculate the contour values corresponding to given HPD confidence
levels for `_fast_kde_2d`.

* Add option to plot HPD regions to `plot_kde`

* Change `hpd_levels` to `levels` in `plot_kde`

This commit refactors the argument `hpd_levels` to `levels` in
`plot_kde`. Additionally, it changes all references to "highest
posterior density" in comments and error messages to "highest density
interval".

* Refactor `hpd_levels` to `hdi_levels` in helper fn

This commit refactors `hpd_levels` to `hdi_levels` in the function
`_find_hpd_contours`. It additionally replaces all mentions of "highest
probability density" with "highest density interval".

* Refactor HDI contour fn to `_find_hdi_contours`

As with the last two commits, this commit removes the remaining
references to "highest posterior density", replacing
`_find_hpd_contours` with `_find_hdi_contours`.

* Update contour-finding to use corner.py algorithm

This commit updates `_find_hdi_contours` to use the method that is
employed in the package `corner.py`.

* Increase clarity in `plot_kde` contour-finding

* Add unit test for `_find_hdi_contours`

* Add unit tests for `levels` keyword of `plot_kde`

* Make minor stylistic changes for pylint compliance

* Update `_find_hdi_contours` docstring

* Update `plot_kde` docstring

* Update new unit tests for mypy compliance

* Refactor `levels` and `hdi_levels` to `hdi_probs`

This commit will refactor the `levels` argument of `plot_kde()` and the
`hdi_levels` argument of `_find_hdi_contours()` to `hdi_probs` for both,
with the hope that this provides an interface consistent with that given
for `arviz.hdi()`.

* Add an example of the HDI contour plot interface

* Add example to `plot_kde` docstring

* Update CHANGELOG.md to add HDI contour feature

* Update `plot_pair` to pass `kde_kwargs` by value

This commit passes a copy of the `kde_kwargs` dict when calling the
`plot_kde` method multiple times. If the dictionary is not copied before
being passed, the dictionary is effectively passed by value, which can
lead to unexpected behaviour when multiple calls to `kde_kwargs` are
made.

* Add example of KDE pair plot with HDI contours

* Update `kde_plot` HDI contour example docstring

* Update docstring for `plot_kde`

* Add tests to catch bug in `plot_kde` with Bokeh

* Update docstring for `hdi_probs` arg in `plot_kde`

* Add unit tests for HDI contours with Bokeh backend

* Update title of existing Bokeh 2d KDE example

This commit changes the title of `examples/bokeh/bokeh_plot_kde_2d.py`
to be in line with the corresponding matplotlib example. (With the
inconsistent titles, the title for both the matplotlib and bokeh
examples showed as "2d KDE Bokeh" on the example gallery main page,
which is misleading).

* Add new examples for 2d KDE plots with Bokeh

* Update Bokeh `plot_pair` to deepcopy `kde_kwargs`

This commit will update `plot_pair` to deepcopy `kde_kwargs`, fixing a
bug that caused unexpected behaviour with the kde formatting.

* Update example 2d KDE plot titles for clarity

* Update `plot_kde` to use `hdi_probs` over `levels`

This commit updates `plot_kde` to use `hdi_probs` over `levels` if both
are specified, and warn the user that both have been specified.

* Add unit tests to test `plot_kde` warnings

This commit adds a couple of new unit tests to verify that the user is
warned when both `levels` and `hdi_probs` have been specified to
`plot_kde`.

* Update code to be Pylint-2.8.1 compliant