Allow specification of log_likelihood variable name in loo_pit
It may not always be the case that variables in the log_likelihood group match exactly the names in observed or posterior predictive groups. This change allows to pass a str as "log_weights" which points to the log_likelihood variable to pair with the observed and posterior predictive ones.
This is helpful with InferenceData coming from Stan, which won't allow, as far as I know, to use the same variable name for observed and log-likelihood variables in the same model file.
I've added the corresponding test.