Add dict to `from_cmdstan`'s log_likelihood parameter accepted types  (#1978)
* Allow specification of log_likelihood variable name in loo_pit

It may not always be the case that variables in the log_likelihood group match exactly the names in observed or posterior predictive groups. This change allows to pass a str as "log_weights" which points to the log_likelihood variable to pair with the observed and posterior predictive ones.
This is helpful with InferenceData coming from Stan, which won't allow, as far as I know, to use the same variable name for observed and log-likelihood variables in the same model file.
I've added the corresponding test.

* Update CHANGELOG.md

* Revert "Update CHANGELOG.md"

This reverts commit 80043f4bfb6afc665371c38dc1cb6a2a8fc6a0ba.

* Revert "Allow specification of log_likelihood variable name in loo_pit"

This reverts commit 7d713db593bcda91f8c006316ee2f59fdcbd5785.

* Add dict to `from_cmdstan`'s log_likelihood parameter accepted types

This works identical as in `from_pystan`, allowing to map observed variable names to their corresponding log-likelihoods.
* Added test.

* Fix column parsing to work with Python <3.8

Dropped the assignment expression. The column name and indices are now split once in a generator to avoid splitting the column string several times.

* Deal with columns with two or more dimensions

* Fix broken logic in the last commit

Columns representing scalar variables are now correctly handled, those representing vectors will still pass their full multi-dimensional indices to the renamed variable.

* Fixed pylint `cell-var-from-loop` warning

pylint caught an issue. The logic to translate variable names got overly complicated, my bad. I have simplified it.